# manager-external-client/api/openapi.yaml
openapi: 3.0.0
info:
  description: Public API for the Preservation Manager.
  title: Preservation Manager Public API
  version: 1.0.0
servers:
- description: Local server
  url: http://localhost:8484/api
security:
- ApiKeyAuth: []
...
paths:
...
  /vault/{orgName}:
    get:
      operationId: listVaults
      parameters:
      - description: The ID of the organization
        example: organization-id
        in: path
        name: orgName
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListVaultsResponse'
          description: List of vaults
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Lists all vaults for an organization
      tags:
      - PreservationManager
      x-accepts: application/json
  /vault/{orgName}/permission:
    get:
      operationId: listOrgUserVaultPermissions
      parameters:
      - description: The ID of the organization
        example: organization-id
        in: path
        name: orgName
        required: true
        schema:
          type: string
      - description: A user's username in the preservation system
        in: query
        name: username
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListVaultPermissionsResponse'
          description: List of all vaults in an org along with the user's permissions
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: List all of the vaults in an org along with the user's permissions
      tags:
      - PreservationManager
      x-accepts: application/json
  /vault/{vault}/user:
    get:
      operationId: listUsersInVault
      parameters:
      - description: The name of the vault
        example: vault-name
        in: path
        name: vault
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListVaultUsersResponse'
          description: List of all of the users who have access to the given vault
            along with their permissions
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Retrieves a list of users with read or write permissions for the given
        vault
      tags:
      - PreservationManager
      x-accepts: application/json
  /vault/{vault}/describe:
    get:
      operationId: describeVault
      parameters:
      - description: The name of the vault
        example: vault-name
        in: path
        name: vault
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DescribeVaultResponse'
          description: Details about the vault and the current user's permissions
            in the vault
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Returns details about the vault and the current user's permissions
        in the vault
      tags:
      - PreservationManager
      x-accepts: application/json
  /vault/{vault}/problems:
    get:
      operationId: listVaultProblems
      parameters:
      - description: The name of the vault
        example: vault-name
        in: path
        name: vault
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListVaultProblemsResponse'
          description: List of objects within the vault for which storage problems
            have been identified
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Returns a list of objects within the vault for which storage problems
        have been identified
      tags:
      - PreservationManager
      x-accepts: application/json
  /vault:
    post:
      operationId: createVault
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVaultRequest'
      responses:
        "200":
          description: Success
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Creates a new vault for an organization
      tags:
      - PreservationManager
      x-contentType: application/json
      x-accepts: application/json
  /vault/retrieve:
    post:
      operationId: retrieveObjects
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetrieveObjectsRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrieveObjectsResponse'
          description: The result of the retrieve request operation
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Requests the retrieval of preserved objects
      tags:
      - PreservationManager
      x-contentType: application/json
      x-accepts: application/json
  /vault/retrieve/{requestId}:
    get:
      operationId: describeRetrieveRequest
      parameters:
      - description: The ID of the request
        example: 10293
        in: path
        name: requestId
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DescribeRetrieveObjectsResponse'
          description: Details about a retrieve objects request
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Describes a retrieve objects request
      tags:
      - PreservationManager
      x-accepts: application/json
  /vault/retrieve/{requestId}/retry/{jobId}:
    post:
      operationId: retryRetrieveJob
      parameters:
      - description: The ID of the request
        example: 10293
        in: path
        name: requestId
        required: true
        schema:
          format: int64
          type: integer
      - description: The ID of the job
        example: 10293
        in: path
        name: jobId
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: Success
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Response returned when an operation fails.
      summary: Retries a failed retrieve job
      tags:
      - PreservationManager
      x-accepts: application/json
  /vault/retrieve/download/{jobId}:
    get:
      operationId: downloadJob
      parameters:
      - description: The ID of the job
        example: 10293
        in: path
        name: jobId
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          content:
            application/zip:
              schema:
                format: binary
                type: string
          description: OK
      summary: Downloads the output of a retrieve job
      tags:
      - PreservationManager
      x-accepts: application/zip
...
components:
  schemas:
...
    RetrieveObjectsRequest:
      example:
        externalObjectIds:
        - 1711.dl/ZQ5WPGU2GKBFT8Q
        - 1711.dl/ZQ5WPGU2GKBFT8Q
        allVersions: true
        vault: vault-name
      properties:
        vault:
          description: The name of the vault
          example: vault-name
          type: string
        allVersions:
          description: true to return every version of each object
          example: true
          type: boolean
        externalObjectIds:
          description: Objects to retrieve
          items:
            $ref: '#/components/schemas/ExternalObjectId'
          type: array
      required:
      - vault
    RetrieveObjectsResponse:
      example:
        requestId: 10293
      properties:
        requestId:
          description: The id of the retrieval request
          example: 10293
          format: int64
          type: integer
      required:
      - requestId
    RetrieveRequestPart:
      example:
        jobId: 10293
        state: READY
        lastDownloadedTimestamp: 2000-01-23T04:56:07.000+00:00
      properties:
        jobId:
          description: The id of a job
          example: 10293
          format: int64
          type: integer
        state:
          description: Indicates the state of this request part
          enum:
          - READY
          - NOT_READY
          - FAILED
          - DELETED
          - CANCELLED
          type: string
        lastDownloadedTimestamp:
          description: Timestamp this part was last downloaded
          format: date-time
          type: string
    DescribeRetrieveObjectsResponse:
      example:
        deleted: true
        externalObjectIds:
        - 1711.dl/ZQ5WPGU2GKBFT8Q
        - 1711.dl/ZQ5WPGU2GKBFT8Q
        allVersions: true
        createdTimestamp: 2000-01-23T04:56:07.000+00:00
        parts:
        - jobId: 10293
          state: READY
          lastDownloadedTimestamp: 2000-01-23T04:56:07.000+00:00
        - jobId: 10293
          state: READY
          lastDownloadedTimestamp: 2000-01-23T04:56:07.000+00:00
        vault: vault-name
        deletedTimestamp: 2000-01-23T04:56:07.000+00:00
        username: username
      properties:
        username:
          description: A user's username in the preservation system
          type: string
        vault:
          description: The name of the vault
          example: vault-name
          type: string
        allVersions:
          description: true to return every version of each object
          example: true
          type: boolean
        externalObjectIds:
          description: Objects to retrieve
          items:
            $ref: '#/components/schemas/ExternalObjectId'
          type: array
        deleted:
          description: Indicates if the DIPs generated by this request have been deleted
          example: true
          type: boolean
        createdTimestamp:
          description: Timestamp the request was received
          format: date-time
          type: string
        deletedTimestamp:
          description: Timestamp the DIPS were deleted
          format: date-time
          type: string
        parts:
          description: The parts this request was divided into
          items:
            $ref: '#/components/schemas/RetrieveRequestPart'
          type: array
...
    ListVaultsResponse:
      example:
        vaults:
        - name: vault-name
          description: Vault description with more details than the vault name
        - name: vault-name
          description: Vault description with more details than the vault name
      properties:
        vaults:
          description: List of vault names and descriptions by organization
          items:
            $ref: '#/components/schemas/VaultObject'
          type: array
    ListVaultPermissionsResponse:
      example:
        vaultPermissions:
        - permissions:
          - null
          - null
          vault:
            name: vault-name
            description: Vault description with more details than the vault name
        - permissions:
          - null
          - null
          vault:
            name: vault-name
            description: Vault description with more details than the vault name
      properties:
        vaultPermissions:
          description: List of all vaults in an org along with the user's permissions
          items:
            $ref: '#/components/schemas/VaultPermissionsWithDescription'
          type: array
    CreateVaultRequest:
      example:
        orgName: organization-id
        vault:
          name: vault-name
          description: Vault description with more details than the vault name
      properties:
        orgName:
          description: The ID of the organization
          example: organization-id
          type: string
        vault:
          $ref: '#/components/schemas/VaultObject'
      required:
      - orgName
      - vault
    DescribeVaultResponse:
      example:
        objectCount: 10293
        orgName: organization-id
        permissions:
        - null
        - null
        vault:
          name: vault-name
          description: Vault description with more details than the vault name
        approximateStorageMegabytes: 10293
      properties:
        orgName:
          description: The ID of the organization
          example: organization-id
          type: string
        vault:
          $ref: '#/components/schemas/VaultObject'
        objectCount:
          description: The number of objects in the vault
          example: 10293
          format: int64
          type: integer
        approximateStorageMegabytes:
          description: The approximate amount of space occupied by all object content
            files in the vault in megabytes
          example: 10293
          format: int64
          type: integer
        permissions:
          description: The permissions the user has in the vault
          items:
            $ref: '#/components/schemas/VaultPermission'
          type: array
    ListVaultProblemsResponse:
      example:
        objects:
        - 1711.dl/ZQ5WPGU2GKBFT8Q
        - 1711.dl/ZQ5WPGU2GKBFT8Q
      properties:
        objects:
          description: List of external objects IDs for objects with storage problems
          items:
            $ref: '#/components/schemas/ExternalObjectId'
          type: array
...
    ListVaultUsersResponse:
      example:
        users:
        - displayName: displayName
          permissions:
          - null
          - null
          userType: GENERAL_USER
          enabledInOrg: true
          username: username
        - displayName: displayName
          permissions:
          - null
          - null
          userType: GENERAL_USER
          enabledInOrg: true
          username: username
      properties:
        users:
          description: List of users in vault
          items:
            $ref: '#/components/schemas/VaultUser'
          type: array
...
    ErrorResponse:
      properties:
        timestamp:
          description: Timestamp when the error occurred
          format: date-time
          type: string
        requestId:
          description: UUID assigned to the request
          example: 6037de0a4b2e4611956da540938a00ff
          type: string
        status:
          description: HTTP status code
          example: 500
          format: int32
          type: integer
        error:
          description: Application classification of the error
          type: string
        message:
          description: Description of the error
          example: Invalid input
          type: string
        path:
          description: Request path
          example: /api/ingest/bag
          type: string
      required:
      - error
      - message
      - path
      - requestId
      - status
      - timestamp
...
    ExternalObjectId:
      description: The external ID of the object
      example: 1711.dl/ZQ5WPGU2GKBFT8Q
      type: string
...
    VaultObject:
      example:
        name: vault-name
        description: Vault description with more details than the vault name
      properties:
        name:
          description: The name of the vault
          example: vault-name
          type: string
        description:
          description: Additional information about the vault
          example: Vault description with more details than the vault name
          type: string
    VaultPermissionsWithDescription:
      example:
        permissions:
        - null
        - null
        vault:
          name: vault-name
          description: Vault description with more details than the vault name
      properties:
        vault:
          $ref: '#/components/schemas/VaultObject'
        permissions:
          description: The permissions the user has in the vault
          items:
            $ref: '#/components/schemas/VaultPermission'
          type: array
    VaultPermission:
      description: A vault permission
      enum:
      - READ
      - WRITE
      type: string
...
    UserType:
      description: Indicates the type of user
      enum:
      - GENERAL_USER
      - PROGRAMMATIC_USER
      - SERVICE_ADMIN
      example: GENERAL_USER
      type: string
...
    VaultUser:
      description: Details about a user in a vault
      example:
        displayName: displayName
        permissions:
        - null
        - null
        userType: GENERAL_USER
        enabledInOrg: true
        username: username
      properties:
        username:
          description: A user's username in the preservation system
          type: string
        displayName:
          description: The user's display name
          type: string
        enabledInOrg:
          description: Indicates if the user is enabled
          example: true
          type: boolean
        userType:
          $ref: '#/components/schemas/UserType'
        permissions:
          description: The permissions the user has in the vault
          items:
            $ref: '#/components/schemas/VaultPermission'
          type: array
      type: object
...
  securitySchemes:
    ApiKeyAuth:
      in: header
      name: X-API-KEY
      type: apiKey